Issue 495 - (WIP) factored out card specific code from _slug into FullCard.vue
